+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Race condition in pin entering that caused double entry
 - Glitch where "next" button would appear at the top of the pin screen
 
+### Internal
+- Upgrade Flutter to 3.44.4 and upgrade all dependencies
+
 ## [8.0.0] - 2026-06-22
 ### Changed
 - Make data tab, credential details, activity, notifications and add-data screens schema-independent